move SAMHparams
